requirements.txt
agora-token-builder==1.0.0
asgiref==3.4.1
cffi==1.15.1
cryptography==37.0.4
Django==4.0.1
gunicorn==20.1.0
mysqlclient==2.1.0
Pillow==9.1.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
sqlparse==0.4.2
tzdata==2021.5
whitenoise==6.1.0
